+++ plutus-context-builder/src/Test/Plutus/ContextBuilder.hs
@@ -25,6 +25,7 @@ module Test.Plutus.ContextBuilder (
   ValueType (..),
   SideUTXO (..),
   ValidatorUTXO (..),
+  SomeValidatedUTXO (..),
   TestUTXO (..),
   ValidatorUTXOs (..),
   Minting (..),
@@ -42,8 +43,10 @@ module Test.Plutus.ContextBuilder (
   -- ** Basic construction
   input,
   validatorInput,
+  validatedInput,
   output,
   validatorOutput,
+  validatedOutput,
   signedWith,
   datum,
   addDatum,
@@ -85,12 +88,15 @@ module Test.Plutus.ContextBuilder (
   -- ** Direct
   liftContextFragment,
   liftNamedContextFragment,
+  foldBuilt,
 
   -- ** Build finished context
   spendingScriptContext,
   mintingScriptContext,
   spendingScriptContextDef,
   mintingScriptContextDef,
+  transactionSpending,
+  transactionMinting,
 
   -- ** Utilities
   defTransactionConfig,
@@ -123,8 +129,9 @@ import Test.Plutus.ContextBuilder.Internal (
   InputPosition (Head, Tail),
   Minting (Mint),
   Naming (Anonymous, Named),
-  Purpose (ForMinting, ForSpending),
+  Purpose (ForMinting, ForSpending, ForTransaction),
   SideUTXO (SideUTXO, sUtxoType, sUtxoValue),
+  SomeValidatedUTXO (SomeValidatedUTXO, someRedeemer, someSpendingScript, someUTxO),
   TestUTXO (TestUTXO, tUtxoDatum, tUtxoValue),
   TransactionConfig (
     TransactionConfig,
@@ -137,14 +144,17 @@ import Test.Plutus.ContextBuilder.Internal (
   ),
   UTXOType (PubKeyUTXO, ScriptUTXO),
   ValidatorUTXO (ValidatorUTXO, vUtxoDatum, vUtxoValue),
-  ValidatorUTXOs (NoValidatorUTXOs, ValidatorUTXOs),
+  ValidatorUTXOs (MultiValidatorUTXOs, NoValidatorUTXOs, ValidatorUTXOs),
   ValueType (GeneralValue, TokensValue),
   defTransactionConfig,
+  foldBuilt,
   makeIncompleteContexts,
   mintingScriptContext,
   mintingScriptContextDef,
   spendingScriptContext,
   spendingScriptContextDef,
+  transactionMinting,
+  transactionSpending,
  )
 import Test.Plutus.ContextBuilder.Minting (
   MintingPolicyAction (BurnAction, MintAction),
@@ -182,7 +192,32 @@ validatorInput ::
   ValidatorUTXO d ->
   ContextBuilder ( 'ForSpending d r) 'Anonymous
 validatorInput name x =
-  NoNames $ mempty {cfValidatorInputs = ValidatorUTXOs $ Map.singleton name x}
+  NoNames $
+    mempty
+      { cfValidatorInputs =
+          ValidatorUTXOs $ Map.singleton name x
+      }
+
+{- | Anonymous context from a single 'SomeValidatedUTXO' input.
+
+ = Note
+
+ This input won't be used for spending in any 'ScriptPurpose' of any
+ 'ScriptContext' built from this.
+
+ @since 2.0.1
+-}
+validatedInput ::
+  -- | Name of the input
+  Text ->
+  SomeValidatedUTXO ->
+  ContextBuilder 'ForTransaction 'Anonymous
+validatedInput name x =
+  NoNames $
+    mempty
+      { cfValidatorInputs =
+          MultiValidatorUTXOs $ Map.singleton name x
+      }
 
 {- | Anonymous context from a single 'SideUTXO' output.
 
@@ -205,7 +240,26 @@ validatorOutput ::
   ValidatorUTXO d ->
   ContextBuilder ( 'ForSpending d r) 'Anonymous
 validatorOutput name x =
-  NoNames $ mempty {cfValidatorOutputs = ValidatorUTXOs $ Map.singleton name x}
+  NoNames $
+    mempty
+      { cfValidatorOutputs =
+          ValidatorUTXOs $ Map.singleton name x
+      }
+
+{- | Anonymous context from a single 'SomeValidatedUTXO' output.
+
+ @since 2.0.1
+-}
+validatedOutput ::
+  Text ->
+  SomeValidatedUTXO ->
+  ContextBuilder 'ForTransaction 'Anonymous
+validatedOutput name x =
+  NoNames $
+    mempty
+      { cfValidatorOutputs =
+          MultiValidatorUTXOs $ Map.singleton name x
+      }
 
 {- | Anonymous context signed with one signature.