+++ tasty-plutus/test/Properties/Validator.hs
@@ -14,7 +14,10 @@ import Test.Plutus.ContextBuilder (
   ContextBuilder,
   Naming (Anonymous),
   Purpose (ForSpending),
+  SomeValidatedUTXO (SomeValidatedUTXO),
+  ValidatorUTXO (ValidatorUTXO),
   outToPubKey,
+  validatedInput,
  )
 import Test.QuickCheck.Plutus.Instances ()
 import Test.Tasty (TestTree, testGroup)
@@ -22,6 +25,10 @@ import Test.Tasty.Plutus.Script.Property (
   paramScriptProperty,
   scriptProperty,
   scriptPropertyPass,
+ )
+import Test.Tasty.Plutus.Script.Unit (
+  shouldValidateTransaction,
+  shouldn'tValidateTransaction,
  )
 import Test.Tasty.Plutus.TestData (
   Generator (GenForSpending),
@@ -77,6 +84,12 @@ tests =
         "Validator checks secret key"
         (\(secret, _, _) -> paramTestValidator secret)
         $ GenForSpending genForParam transformForParam
+    , shouldValidateTransaction "Unit transaction test" mempty $
+        validatedInput "input" $
+          SomeValidatedUTXO (ValidatorUTXO (1, 2) mempty) simpleTestValidator (3, 2)
+    , shouldn'tValidateTransaction "Negative unit transaction test" mempty $
+        validatedInput "input" $
+          SomeValidatedUTXO (ValidatorUTXO (1, 2) mempty) simpleTestValidator (3, 4)
     ]
 
 genForSimple :: Methodology (Integer, Integer, Integer, Integer, Value)